teaching_panel/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
PyJWT>=2.8.0
requests>=2.31.0
python-decouple>=3.8
python-dotenv>=1.0.0
djangorestframework-simplejwt>=5.3.0
celery>=5.3.0
redis>=5.0.0
azure-cosmos>=4.7.0
django-celery-beat>=2.8.0
django-recaptcha>=4.1.0
